[BUGFIX] Keep int as a unsigned long in

+++ 0x08-palindrome_integer/0-is_palindrome.c
@@ -7,7 +7,7 @@
  * @n: get max tenth of this number
  * Return: tenth
  */
-int get_tenth(unsigned long n)
+unsigned long int get_tenth(unsigned long n)
 {
 	unsigned long i;
